+++ NEWS.md
@@ -1,5 +1,7 @@
 # jlme (development version)
 
+* Experimental support for `MixedModels.parametricbootstrap()` via `parametricbootstrap()` with a `tidy()` method.
+
 * `jlme_setup()` gains an `add` argument which accepts a character vector of additional Julia libraries to install and load on initializing the session.
 
 * Some common workflows in MixedModels.jl have been re-exported as functions: `propertynames()`, `issingular()`